+++ src/app/matching/page.tsx
@@ -1,3 +1,19 @@
+import InterestStep from "@/components/domains/matching/steps/InterestStep";
+import LevelStep from "@/components/domains/matching/steps/LevelStep";
+import LocationStep from "@/components/domains/matching/steps/LocationStep";
+import MoodStep from "@/components/domains/matching/steps/MoodStep";
+import SuccessStep from "@/components/domains/matching/steps/SuccessStep";
+import TypeStep from "@/components/domains/matching/steps/TypeStep";
+
 export default function MatchingPage() {
-  return null;
+  return (
+    <>
+      <InterestStep />
+      {/* <LevelStep /> */}
+      {/* <LocationStep /> */}
+      {/* <MoodStep /> */}
+      {/* <TypeStep /> */}
+      {/* <SuccessStep /> */}
+    </>
+  );
 }
