+++ src/components/commons/SelectLocation.tsx
@@ -35,63 +35,61 @@ export default function SelectLocation() {
   };
 
   return (
-    <>
-      <div className="flex flex-col gap-[40px] border-y border-gray-300 h-[333px]">
-        <div className="flex justify-between items-center h-full">
-          <div className="flex flex-col w-[117px] h-full border-r border-gray-300 scrollbar-hide overflow-x-scroll">
-            {Object.keys(content).map((key) => (
-              <button
-                key={key}
-                type="button"
-                className={`w-full py-[14px] px-[10px] text-center text-gray-600 font-medium tracking-[-0.28px] leading-[21px] bg-gray-100 ${
-                  selectedCity === key ? "bg-main-500 text-white" : ""
-                }`}
-                onClick={() => handleCityClick(key)}>
-                {key}
-              </button>
+    <div className="flex flex-col gap-[20px]">
+      <div className="flex justify-between items-center h-[333px] border-y border-gray-300">
+        <div className="flex flex-col w-[117px] h-full border-r border-gray-300 scrollbar-hide overflow-x-scroll">
+          {Object.keys(content).map((key) => (
+            <button
+              key={key}
+              type="button"
+              className={`w-full py-[14px] px-[10px] text-center text-gray-600 font-medium tracking-[-0.28px] leading-[21px] bg-gray-100 ${
+                selectedCity === key ? "bg-main-500 text-white" : ""
+              }`}
+              onClick={() => handleCityClick(key)}>
+              {key}
+            </button>
+          ))}
+        </div>
+        <div className="flex flex-col w-full h-full scrollbar-hide overflow-x-scroll">
+          {selectedCity && (
+            <div className="flex flex-col">
+              {content[selectedCity as keyof CityList].map((district) => {
+                const isSelected = selectedLocations.some(
+                  (location) => location.city === selectedCity && location.district === district,
+                );
+                return (
+                  <button
+                    key={`${selectedCity}-${district}`}
+                    type="button"
+                    className={`flex justify-between py-[14px] px-5 w-full text-center text-gray-600 font-medium tracking-[-0.28px] leading-[21px] border-b-[1px] border-gray-300 ${
+                      isSelected ? "bg-main-100 text-main-500" : ""
+                    }`}
+                    onClick={() => handleDistrictClick(selectedCity, district)}>
+                    {district}
+                    <Image src={isSelected ? checkBlue : checkGray} alt="체크 아이콘" width={24} height={24} />
+                  </button>
+                );
+              })}
+            </div>
+          )}
+        </div>
+      </div>
+      {selectedLocations.length > 0 && (
+        <>
+          <div className="flex items-center gap-[11px] sm:gap-[14px]">
+            {selectedLocations.map((location) => (
+              <div
+                key={`${location.city}-${location.district}`}
+                className="flex justify-center items-center h-10 py-[9px] px-[3px] sm:px-[14px] gap-2 bg-main-100 border border-main-500 rounded-[8px] text-main-500 text-[14px] font-medium tracking-[-0.28px] leading-[21px]">
+                {location.city} {location.district}
+                <button type="button" onClick={() => handleRemoveLocation(location.district)}>
+                  <Image src={closeBlue} alt="닫기 아이콘" width={18} height={18} />
+                </button>
+              </div>
             ))}
           </div>
-          <div className="flex flex-col w-full h-full scrollbar-hide overflow-x-scroll">
-            {selectedCity && (
-              <div className="flex flex-col">
-                {content[selectedCity as keyof CityList].map((district) => {
-                  const isSelected = selectedLocations.some(
-                    (location) => location.city === selectedCity && location.district === district,
-                  );
-                  return (
-                    <button
-                      key={`${selectedCity}-${district}`}
-                      type="button"
-                      className={`flex justify-between py-[14px] px-5 w-full text-center text-gray-600 font-medium tracking-[-0.28px] leading-[21px] border-b-[1px] border-gray-300 ${
-                        isSelected ? "bg-main-100 text-main-500" : ""
-                      }`}
-                      onClick={() => handleDistrictClick(selectedCity, district)}>
-                      {district}
-                      <Image src={isSelected ? checkBlue : checkGray} alt="체크 아이콘" width={24} height={24} />
-                    </button>
-                  );
-                })}
-              </div>
-            )}
-          </div>
-        </div>
-        <ul className="grid sm:grid-cols-3 grid-cols-2 items-center gap-[14px] h-[39px] w-full">
-          {selectedLocations.length > 0 && (
-            <>
-              {selectedLocations.map((location) => (
-                <li
-                  key={`${location.city}-${location.district}`}
-                  className="flex justify-between items-center py-[9px] px-[14px] w-full h-[40px] bg-main-100 border border-main-500 rounded-[8px] text-main-500 text-center text-[14px] font-medium tracking-[-0.28px] leading-[21px]">
-                  {location.city} {location.district}
-                  <button type="button" onClick={() => handleRemoveLocation(location.district)}>
-                    <Image src={closeBlue} alt="닫기 아이콘" width={18} height={18} />
-                  </button>
-                </li>
-              ))}
-            </>
-          )}
-        </ul>
-      </div>
-    </>
+        </>
+      )}
+    </div>
   );
 }
